Pushing bsa archive types back together into one file
