+++ docs/source/bethesda_structs.archive.rst
@@ -1,0 +1,16 @@
+bethesda\_structs\.archive
+==========================
+
+.. automodule:: bethesda_structs.archive
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+
+bsa
+---
+
+.. automodule:: bethesda_structs.archive.bsa
+   :members:
+   :undoc-members:
+   :show-inheritance:
